[domain] move LocalFile into module
