+++ domain/src/main/scala/net/kemitix/s3thorp/domain/LocalFile.scala
@@ -1,9 +1,7 @@
-package net.kemitix.s3thorp
+package net.kemitix.s3thorp.domain
 
 import java.io.File
 import java.nio.file.Path
-
-import net.kemitix.s3thorp.domain.{MD5Hash, RemoteKey}
 
 final case class LocalFile(
   file: File,
