+++ src/main/scala/net/kemitix/s3thorp/LocalFileStream.scala
@@ -2,7 +2,7 @@ package net.kemitix.s3thorp
 
 import java.io.File
 
-import net.kemitix.s3thorp.domain.MD5Hash
+import net.kemitix.s3thorp.domain.{LocalFile, MD5Hash}
 
 class LocalFileStream(md5HashGenerator: File => MD5Hash)
   extends KeyGenerator
