+++ src/main/scala/net/kemitix/s3thorp/S3MetaData.scala
@@ -1,5 +1,7 @@
 package net.kemitix.s3thorp
 
+import net.kemitix.s3thorp.domain.LocalFile
+
 // For the LocalFile, the set of matching S3 objects with the same MD5Hash, and any S3 object with the same remote key
 final case class S3MetaData(
   localFile: LocalFile,
