+++ src/main/scala/net/kemitix/s3thorp/awssdk/ThorpS3Client.scala
@@ -5,7 +5,7 @@ import com.amazonaws.services.s3.AmazonS3
 import com.amazonaws.services.s3.transfer.TransferManager
 import com.github.j5ik2o.reactive.aws.s3.cats.S3CatsIOClient
 import net.kemitix.s3thorp._
-import net.kemitix.s3thorp.domain.{Bucket, MD5Hash, RemoteKey}
+import net.kemitix.s3thorp.domain.{Bucket, LocalFile, MD5Hash, RemoteKey}
 import software.amazon.awssdk.services.s3.model.{Bucket => _}
 
 class ThorpS3Client(ioS3Client: S3CatsIOClient,
